requirements.txt
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
more-itertools==10.7.0
pycparser==2.22
PyGithub==2.7.0
PyJWT==2.10.1
PyNaCl==1.5.0
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
